+++ Code/Data/userSettings.json
@@ -1,53 +1,39 @@
 {
-  "userSettings": [
-    {
-      "username": "user1",
-      "language": "english",
-      "email": "True",
-      "SMS": "False",
-      "ads": "True"
-    },
-    {
-      "username": "user2",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    },
-    {
-      "username": "testuser1",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    },
-    {
-      "username": "testuser2",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    },
-    {
-      "username": "test",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    },
-    {
-      "username": "test",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    },
-    {
-      "username": "test",
-      "language": "english",
-      "email": "True",
-      "SMS": "True",
-      "ads": "True"
-    }
-  ]
+    "userSettings": [
+        {
+            "username": "user1",
+            "language": "english",
+            "email": "False",
+            "SMS": "False",
+            "ads": "False"
+        },
+        {
+            "username": "user2",
+            "language": "english",
+            "email": "True",
+            "SMS": "True",
+            "ads": "True"
+        },
+        {
+            "username": "testuser1",
+            "language": "english",
+            "email": "True",
+            "SMS": "True",
+            "ads": "True"
+        },
+        {
+            "username": "testuser2",
+            "language": "english",
+            "email": "True",
+            "SMS": "True",
+            "ads": "True"
+        },
+        {
+            "username": "test",
+            "language": "english",
+            "email": "True",
+            "SMS": "False",
+            "ads": "False"
+        }
+    ]
 }
