+++ Code/Source/homePageOptions.py
@@ -1,0 +1,131 @@
+import json
+from Code.Source.globalVariables import addPage, getFirst, getJobFile, getLast, getLoggedUser
+from Code.Source.menuOptions import back
+from Code.Source.utility import printDivider, writeJson
+
+
+def showHomePageGreeting():
+    printDivider()
+    print("Welcome to InCollege!")
+    print("""Please choose from one of the options below:\n1. Search for a job
+2. Find someone you know\n3. Learn a new skill\n4. Useful Links\n5. InCollege Important Links\n6. Go to previously visited page\n""")
+
+def showSkillPageGreeting():
+    printDivider()
+    print("""Learn and explore the skill options below!\n1. Communication\n2. Leadership\n3. Collaboration\n4. Responsibility\n5. Time Management\n
+Please enter a number from 1-5.\nEnter x to return to the home page.\nEnter y to go to previously visited page.\n""")
+
+def showConstructionMessage(message):
+    printDivider()
+    print("We're sorry!\n'{}' feature is still under construction.\n".format(message))
+
+def returnToHomePage():
+    if getLoggedUser() == None:
+        return
+    user_choice = input('Go back to homepage? (y/n): ')
+    while user_choice != 'y' and user_choice != 'n':
+        user_choice = input('Please enter y for yes, n for no: ')
+    if user_choice == 'y':
+        return "homePage"
+    else: 
+        printDivider()
+        exit
+
+def jobPage():
+    addPage(jobPage)
+
+    message = "\n1. Post a job\n2. Home page\n3. Previous Page\n"
+    print(message)
+    user_choice = input("Enter your option: ")
+    while user_choice != '1' and user_choice != '2' and user_choice != '3':
+        print("\nInvalid input.\n" + message)
+        user_choice = input("Enter your option: ")
+    if user_choice == '1':
+        addJobPost()
+    elif user_choice == '2':
+        return "homePage"
+    elif user_choice == '3':
+        back()
+
+def addJobPost():
+    fileName = getJobFile()
+
+    #Checks if already 5 job posts
+    with open (fileName) as jsonFile:
+        data = json.load(jsonFile)
+        temp1 = data["numPosts"]
+        if temp1 >= 5:
+            print("There are already five job posts. Try again later.")
+            return
+
+    print("Please input the following information about the job when prompted.\n")
+    while True:
+        jobTitle = input("Job Title: ")
+        length = checkLength(jobTitle, 50)
+        if length == True:
+            break
+    while True:
+        jobDescription = input("Job Description: ")
+        length = checkLength(jobDescription, 250)
+        if length == True:
+            break
+    while True:
+        jobEmployer = input("Employer: ")
+        length = checkLength(jobEmployer, 50)
+        if length == True:
+            break
+    while True:
+        jobLocation = input("Job Location: ")
+        length = checkLength(jobLocation, 50)
+        if length == True:
+            break
+    while True:
+        jobSalary = input("Job Salary: ")
+        length = checkLength(jobSalary, 50)
+        if length == True:
+            break
+    print("\n")
+
+    jobDictionary = {
+        "Title" : jobTitle,
+        "Description" : jobDescription,
+        "Employer" : jobEmployer,
+        "Location" : jobLocation,
+        "Salary" : jobSalary,
+        "Name" : getFirst() + ' ' + getLast()
+    }
+
+    #Appends new post to json file, Increase post count if < 5
+    with open (fileName) as jsonFile:
+        data = json.load(jsonFile)
+        temp = data["jobPosts"]
+        y = jobDictionary
+        temp.append(y)
+
+        #increment number of job posts
+        temp1 = data["numPosts"]
+        data["numPosts"] = temp1 + 1
+    
+    writeJson(data, fileName)
+
+#Read in job posts at application start upclear
+def readJobPosts():
+    fileName = getJobFile()
+    with open(fileName) as json_file:
+        data = json.load(json_file)
+        jobPosts = data["jobPosts"]
+    return jobPosts
+
+#Character Limiter Function (Security Measure)
+def checkLength(input, limit):
+    if len(input) > limit:
+        print("\nERROR: Maximum characters of " + str(limit) + " reached.\n")
+        return False
+    return True
+
+def findSomeonePage():
+    addPage(findSomeonePage)
+    showConstructionMessage("Find someone you know")
+
+def skillPage(skill):
+    showConstructionMessage("Learn a new skill")
