--- Code/Source/successStory.py
@@ -1,8 +1,0 @@
-def storyDisplay():
-    print("'Hello, my name is Jane Doe and let me tell you how InCollege has paved a way to success for me.")
-    print("Not too long ago, I was freshly graduated student that was having a tough time finding jobs and a direction in life.")
-    print("As a freshly graduated student, I didn't have many accomplishments or connections that would help me land job interview,")
-    print("but with the tools and connections that InCollege offered, i'm now well off as an independent adult with a strong career path in software engineering.")
-    print("Not only am I a software engineer, i'm also now an ex-Navy SEAL, an ex-astronaught, and a part time neurosurgeon!")
-    print("All of my accomplishments was stemmed off of me using InCollege. If you too sign up for InCollege, you can also be successful like me!")
-    print("Thanks InCollege!'\n")
