--- Code/Source/writeJson.py
@@ -1,5 +1,0 @@
-import json
-
-def wJson(data, dataFile):
-    with open(dataFile, "w") as json_file:
-        json.dump(data, json_file, indent = 2)
